requirements.txt
qdrant-client
sentence-transformers
google-generativeai
transformers
datasets
pandas
gradio
evaluate
nltk
tqdm
python-dotenv